feat: add indicator to completion
The color of the indicator is based on time since the last change
in the completion. The indicator is green if the last change was
less than 7 minutes ago, yellow if it was between 7 and 15 minutes
ago, and red if it was more than 15 minutes ago.
The indicator is shown in the completion column.
The indicator also has different characters for different
completion types:
- green: "^"
- yellow: "="
- red: "x"